ScreenManager: Increase performance
Increase MAX_FRAME_SKIP and remove gamespeed drop.

before: The gamespeed started to drop at aprox. 30 FPS and the gamespeed
used to drop rapidly bellow 20 FPS.

now: When the FPS is low, the video lags, but the gamespeed is still
constant. It starts to drop slowly when it's bellow 18 FPS.